Remove RxJava dependency
Previous design depends on RxJava.
While reactive extensions are useful,
it is a bit of over-engineering since
I've only included it for the purpose of
multi-threading. I've decided
to drop RxJava previously included in this template
to allow for more flexibility. You can add
RxJava on your projects in needed anyway.